E5/test/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
jinja2==3.1.2
qrcode[pil]==7.4.2
pandas==2.1.4
openpyxl==3.1.2
pillow==10.1.0
reportlab==4.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0 